% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerarModelo.R
\name{entrenar_modelo}
\alias{entrenar_modelo}
\title{Función entrenar_modelo()}
\usage{
entrenar_modelo(df, config)
}
\arguments{
\item{df}{}

\item{config}{}
}
\value{
fit_modelo. El modelo entrenado para df_train
}
\description{
Esta función llama a la función "separa_train()" y sobre el df_train
que devuelve calcula un modelo de regresión lineal para el target basado en los
parámetros fem_particip, gdp y life_ex
Devuelve como salida el modelo entrenado
}
