% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generar_output.R
\name{generarOutput}
\alias{generarOutput}
\title{Funci贸n generarOutput()}
\usage{
generarOutput(modelo, prediccion)
}
\arguments{
\item{modelo}{}

\item{prediccion}{}
}
\value{
prediccion.csv    modelo.rds
}
\description{
Esta funci贸n empaqueta en dos ficheros diferentes los resultados de la ejecuci贸n
del programa. Por un lado, guarda en un .csv los resultados de la predicci贸n.
En un fichero rds guarda el modelo entrenado.
}
