% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerarModelo.R
\name{predecir_objetivo}
\alias{predecir_objetivo}
\title{Función predecir_objetivo()}
\usage{
predecir_objetivo(df, modelo, config)
}
\arguments{
\item{df}{}

\item{modelo}{}

\item{config}{}
}
\value{
predictions La predicción con este modelo para el registro objetivo
}
\description{
Función que predice el target para el registro objetivo 
empleando el modelo entrenado en la función "entrenar_modelo()". 
Para ello, la función llama inicialmene a la función 
"separa_test()" para obtener el registro objetivo
}
