% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerarModelo.R
\name{separa_test}
\alias{separa_test}
\title{Función separa_test()}
\usage{
separa_test(df, config)
}
\arguments{
\item{df}{}

\item{config}{}
}
\value{
df_test
}
\description{
Esta función busca en el df original el registro que combina
el país y anio introducidos por el usuario en el fichero de configuración.
El registro se extrae y se elimina el campo de target y se guarda como 
df_test. Este df se usará en la predicción del target.
}
