% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerarModelo.R
\name{separa_train}
\alias{separa_train}
\title{Función separa_train()}
\usage{
separa_train(df, config)
}
\arguments{
\item{df}{}

\item{config}{}
}
\value{
df_train
}
\description{
Esta función toma el df y busca el registro que combina el dato de 
país y anio que se indican en el fichero config. 
Se crea un nuevo df (df_train) eliminando este registro que servirá
para entrenar el modelo
}
